# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "myProperty API",
        "description": "developed by Jonathan",
        "contact": {
            "email": "jonathanx56@hotmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:80",
    "basePath": "",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "category"
                ],
                "summary": "List Categories",
                "operationId": "list",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "category"
                ],
                "summary": "Create category",
                "operationId": "create",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "-   Enter the name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "-   Enter the description",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "slug",
                        "in": "query",
                        "description": "-   Enter the slug",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            }
        },
        "/api/v1/categories/{id}/real-states": {
            "get": {
                "tags": [
                    "category"
                ],
                "summary": "List Real-state by category",
                "operationId": "list",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "-   Enter the id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            }
        },
        "/api/v1/categories/{id}": {
            "get": {
                "tags": [
                    "category"
                ],
                "summary": "Show Category",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "-   Enter the id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "category"
                ],
                "summary": "Update category",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "-   Enter the id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "-   Enter the name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "-   Enter the description",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "slug",
                        "in": "query",
                        "description": "-   Enter the slug",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "category"
                ],
                "summary": "Delete Category",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "-   Enter the id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "default": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}